move principle mapping into separate pkg
