+++ internal/principals/mapper.go
@@ -1,0 +1,6 @@
+package principals
+
+// PrincipalMapper defines the interface for mapping claims to principals.
+type PrincipalMapper interface {
+	Map(claims interface{}) ([]string, error)
+}
